add amoun in payments

+++ server/app/models.py
@@ -1,6 +1,8 @@
 """SQLModel database models"""
 from datetime import datetime
+from decimal import Decimal
 
+from sqlalchemy import Numeric
 from sqlmodel import Column, DateTime, Field, Integer, SQLModel, String
 
 
@@ -40,7 +42,9 @@ class PaymentLog(SQLModel, table=True):
 
     id: int | None = Field(default=None, primary_key=True)
     member_id: int = Field(sa_column=Column(Integer, nullable=False, index=True))
-    amount: float = Field(sa_column=Column(Integer, nullable=False))  # in cents
+    # Amount in dollars, stored as DECIMAL(10,2) for precise monetary values
+    # Constraints: > 0 and <= 1000, validated at API level
+    amount: Decimal = Field(sa_column=Column(Numeric(10, 2), nullable=False))
     timestamp: datetime = Field(
         default_factory=datetime.utcnow,
         sa_column=Column(DateTime, nullable=False, default=datetime.utcnow),
